fix: Architecture compliance & CI improvements
- Fix 2 file size violations (DIContainer.swift, PayslipParsingProgressService.swift)
- Extract ParsingProgressState enum to separate file for 300-line compliance
- Add comprehensive iOS CI workflow (ios-ci.yml) with:
  - Build & Test with code coverage
  - SPM dependency caching
  - SwiftLint integration
  - Security checks for hardcoded secrets
  - Architecture quality gates
  - PR validation summary
- Update architecture-quality.yml to manual trigger (superseded by ios-ci.yml)
- Remove empty 'workflows 2' directory

Architecture violations: 0
File compliance: 99.7% -> 100%
