+++ PayslipMax/Services/Processing/PayslipParsingProgressService.swift
@@ -9,69 +9,6 @@ import Foundation
 import UIKit
 import Combine
 
-/// Progress state for payslip parsing operations
-enum ParsingProgressState: Equatable {
-    case idle
-    case preparing                          // Setting up for parsing
-    case extracting                         // First LLM pass or OCR
-    case validating                         // Sanity checks
-    case verifying                          // Second LLM pass (if needed)
-    case saving                             // Saving to database
-    case completed(PayslipItem)             // Done with result
-    case failed(String)                     // Error occurred
-
-    var isActive: Bool {
-        switch self {
-        case .idle, .completed, .failed:
-            return false
-        default:
-            return true
-        }
-    }
-
-    var progressMessage: String {
-        switch self {
-        case .idle:
-            return ""
-        case .preparing:
-            return "Preparing payslip..."
-        case .extracting:
-            return "Analyzing payslip... 🔍"
-        case .validating:
-            return "Validating data..."
-        case .verifying:
-            return "Verifying accuracy..."
-        case .saving:
-            return "Saving payslip..."
-        case .completed:
-            return "Complete ✓"
-        case .failed(let error):
-            return "Failed: \(error)"
-        }
-    }
-
-    var progressPercent: Double {
-        switch self {
-        case .idle:
-            return 0.0
-        case .preparing:
-            return 0.1
-        case .extracting:
-            return 0.3
-        case .validating:
-            return 0.6
-        case .verifying:
-            return 0.8
-        case .saving:
-            return 0.95
-        case .completed:
-            return 1.0
-        case .failed:
-            return 0.0
-        }
-    }
-}
-
 /// Service for managing async payslip parsing with progress updates
 @MainActor
 final class PayslipParsingProgressService: ObservableObject {
@@ -95,46 +32,27 @@ final class PayslipParsingProgressService: ObservableObject {
     // MARK: - Public API
 
     /// Starts async parsing with progress updates
-    /// - Parameters:
-    ///   - image: The cropped payslip image
-    ///   - processor: The image import processor
-    func startParsing(
-        image: UIImage,
-        processor: ImageImportProcessor
-    ) {
-        // Cancel any existing task
+    func startParsing(image: UIImage, processor: ImageImportProcessor) {
         currentTask?.cancel()
-
-        // Reset state
         state = .preparing
         hasNewPayslip = false
 
-        // Start async parsing
         currentTask = Task { @MainActor in
             await performParsing(image: image, processor: processor)
         }
     }
 
     /// Starts async parsing with both original and cropped images
-    /// - Parameters:
-    ///   - originalImage: The uncropped original image (for PDF storage)
-    ///   - croppedImage: The cropped image (for LLM processing)
-    ///   - imageIdentifier: UUID for linking to saved image files
-    ///   - processor: The image import processor
     func startParsing(
         originalImage: UIImage,
         croppedImage: UIImage,
         imageIdentifier: UUID?,
         processor: ImageImportProcessor
     ) {
-        // Cancel any existing task
         currentTask?.cancel()
-
-        // Reset state
         state = .preparing
         hasNewPayslip = false
 
-        // Start async parsing with both images
         currentTask = Task { @MainActor in
             await performParsingWithBothImages(
                 originalImage: originalImage,
@@ -162,8 +80,6 @@ final class PayslipParsingProgressService: ObservableObject {
     /// Retries the last failed parsing operation
     func retry() {
         guard let callback = retryCallback else { return }
-
-        // Reset state and retry
         state = .preparing
         hasNewPayslip = false
 
@@ -174,34 +90,22 @@ final class PayslipParsingProgressService: ObservableObject {
 
     // MARK: - Private Methods
 
-    private func performParsing(
-        image: UIImage,
-        processor: ImageImportProcessor
-    ) async {
-        // Store retry callback
+    private func performParsing(image: UIImage, processor: ImageImportProcessor) async {
         retryCallback = { [weak self] in
             await self?.performParsing(image: image, processor: processor)
         }
 
         do {
-            // Step 1: Preparing (brief visual feedback)
             state = .preparing
-
-            // Step 2: Extracting (LLM Pass 1)
             state = .extracting
-
-            // Process the image (this will trigger LLM vision parsing)
-            // Validation and verification happen inside the parser
             let result = await processor.processCroppedImageLLMOnly(image)
 
             switch result {
             case .success(let payslip):
-                // Step 3: Completed with real payslip data
                 state = .completed(payslip)
                 hasNewPayslip = true
-                retryCallback = nil  // Clear retry on success
+                retryCallback = nil
 
-                // Auto-reset after 2 seconds
                 try await Task.sleep(nanoseconds: 2_000_000_000)
                 if case .completed = state {
                     state = .idle
@@ -209,9 +113,7 @@ final class PayslipParsingProgressService: ObservableObject {
 
             case .failure(let error):
                 state = .failed(error.localizedDescription)
-                // Keep retryCallback for retry button
 
-                // Auto-reset after 5 seconds (longer for retry)
                 try await Task.sleep(nanoseconds: 5_000_000_000)
                 if case .failed = state {
                     state = .idle
@@ -220,28 +122,19 @@ final class PayslipParsingProgressService: ObservableObject {
             }
 
         } catch is CancellationError {
-            // Task was cancelled
             state = .idle
             retryCallback = nil
         } catch {
             state = .failed(error.localizedDescription)
-            // Keep retryCallback for retry button
         }
     }
 
-    /// Performs parsing with both original and cropped images
-    /// - Parameters:
-    ///   - originalImage: The uncropped original image (for PDF storage)
-    ///   - croppedImage: The cropped image (for LLM processing)
-    ///   - imageIdentifier: UUID for linking to saved image files
-    ///   - processor: The image import processor
     private func performParsingWithBothImages(
         originalImage: UIImage,
         croppedImage: UIImage,
         imageIdentifier: UUID?,
         processor: ImageImportProcessor
     ) async {
-        // Store retry callback
         retryCallback = { [weak self] in
             await self?.performParsingWithBothImages(
                 originalImage: originalImage,
@@ -252,15 +145,9 @@ final class PayslipParsingProgressService: ObservableObject {
         }
 
         do {
-            // Step 1: Preparing (brief visual feedback)
             state = .preparing
-
-            // Step 2: Extracting (LLM Pass 1)
             state = .extracting
 
-            // Process with BOTH images
-            // Original image -> PDF for storage
-            // Cropped image -> LLM/OCR processing
             let result = await processor.processBothImages(
                 original: originalImage,
                 cropped: croppedImage,
@@ -269,12 +156,10 @@ final class PayslipParsingProgressService: ObservableObject {
 
             switch result {
             case .success(let payslip):
-
                 state = .completed(payslip)
                 hasNewPayslip = true
                 retryCallback = nil
 
-                // Auto-reset after 2 seconds
                 try await Task.sleep(nanoseconds: 2_000_000_000)
                 if case .completed = state {
                     state = .idle
@@ -282,9 +167,7 @@ final class PayslipParsingProgressService: ObservableObject {
 
             case .failure(let error):
                 state = .failed(error.localizedDescription)
-                // Keep retryCallback for retry button
 
-                // Auto-reset after 5 seconds (longer for retry)
                 try await Task.sleep(nanoseconds: 5_000_000_000)
                 if case .failed = state {
                     state = .idle
@@ -293,12 +176,10 @@ final class PayslipParsingProgressService: ObservableObject {
             }
 
         } catch is CancellationError {
-            // Task was cancelled
             state = .idle
             retryCallback = nil
         } catch {
             state = .failed(error.localizedDescription)
-            // Keep retryCallback for retry button
         }
     }
 }
